Move namespaces to submodules
